Have distinct javax+jakarta proxy implementations
